refactored Content and Panel componets.
